backend/requirements.txt
fastapi==0.125.0
uvicorn>=0.29,<1.0
python-multipart>=0.0.9,<1.0
greenlet>=3.0,<4.0
sqlalchemy>=1.4,<2.0
pydantic>=2.13,<3.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
pdfplumber>=0.11,<1.0
google-genai>=1.73,<2.0
python-dotenv>=1.0,<2.0
